Export newRespIOErr
It's otherwise impossible for a 3rd party library to create an IOErr.
I'm using this in order to create a custom pool class that will retry if a
pooled connection returns an IOErr. Right now it's only possible to
return an AppErr (if the dial fails), which makes it harder to distinguish between an actual fatal error.