Defer connection closure outside critical section
When there is a redis node that is stopped/crash, the closure of the
connection can take longer than usual. As the lock on the connection
was taken, the client cannot send any message to the cluster during
this time.

While doing chaos testing, I observed hundreds of milliseconds delay
to send the next message. To prevent that, the closure of the
connection is defered to after the critical section.