refactor PersistentPubSub
The impetus for this change was issue #184, but after reviewing the code
I decided it needed a larger refactor just out of principal. The
original code was written to not use a central go-routine out of a
desire for simplicity, but in the end it ended up still needing
go-routines, and the interacting between them and the locks was anything
but simple.

The new code has a simple go-routine loop which acts in an RPC pattern,
while also monitoring the underlying PubSubConn. It's much easier to
think about than the old code, and not as buggy.