refactor how Pipeline works so that it doesn't necessarily need CmdAction
The new way takes advantage of the new EncodeDecode method and so
doesn't rely on being able to marshal/unmarshal each CmdAction on its
own. This is a step towards one day hopefully getting rid of CmdAction
entirely.

This commit necessarily messes with pipeliner somewhat, and breaks its
behavior in some failure cases. I left a TODO to go back and fix those,
but it's possible pipeliner might be getting significantly refactored
anyway so I didn't spend a lot of time on it.