Implement PipeliningConn and refactor Pipeline
This comes as a response to the change to Conn to make it use
EncodeDecode. The new Pipeline can take in Actions rather than
CmdActions, and the PipeliningConn acts just like a Conn except that in
the background it does the same thing that the old pipeliner used to do.

The old pipeliner is removed, as is all implicit pipelining from the
default pool. Instead it is recommended that the user create their own
Pool using PipeliningConns and differentiate themselves between Actions
they want implicitly pipelined and Actions they don't. Otherwise it is
too difficult for the driver to know what can be pipelined and what
can't (the v3 behavior isn't quite correct).