Add a Tuple type which can be used to capture results from commands like EXEC and EVAL
Without this type it's rather difficult to unmarshal the results from
commands where the result is an array of potentially disparate types.
Without implementing the logic presented in Tuple, the next best way is
to unmarshal into a `[]resp2.RawMessage` and then unmarshal each element
individually. This is neither obvious, pleasant, nor efficient.