Add resp3.Flatten and remove the corresponding functionality from Any
The original idea behind the MarshalBlobString and MarshalNoAggHeaders
was that if Any is having to traverse through the full data structure
anyway then it might as well be able to do the flattening itself as
well. This would save code and be more resource efficient.

In practice this didn't work out. For starters it was necessary to have
a NumElems method, so traversing had to be done twice no matter what.
And when benchmarked doing a Flatten+BlobStringBytes.MarshalRESPs ended
up being faster than NumElems+Any.MarshalRESP, even though Flatten
doesn't interact with the bytes pool. This is likely because it's only
traversing the structure once now, which is much slower than memory
allocations due to the reflection involved.

There is now some duplicated-ish code, but the API is simpler, so I
think overall it's a net win.