Add contexts to actions and EncodeDecode, and refactor the general framework of all components
Somehow these two changes got rolled into one, mainly because adding
Context to EncodeDecode required rewriting Conn which led to a
generalized framework for all components. The framework, proc, primarily
deals with managing the tracking and cleanup of resources opened by the
component, and aids in shutting each component down gracefully.

This fixed a lot of bugs but added new ones. Notably Sentinel is not
built properly, and will deadlock a lot with the new proc. Also
PipeliningConn is in a weird state, and might be superceded by Conn
all-together.