Rewrite Conn, Pool, Stub, get rid of PipeliningConn
The new Conn is essentially a PipeliningConn in its own right, and the
new Pool takes advantage of the Conn in order to share actions across
the pool when possible. For actions which cannot share a Conn Pool will
take a Conn out of the pool until the action is done.

The new Conn and Pool have all new tests, and as part of that Stub was
completely rewritten to match Conn's new behavior.