redesign PubSubConn
The new PubSubConn gets rid of the usage of message channels, as that
caused complications and weird edge-cases, especially when attempting to
close the PubSubConn. If users desire to use channels for consuming
pubsub messages they can still do so, but it's on them to forward the
message from the `Next` method to the channel.

The naming of PubSub related types has also been updated to be more
consistent.

This also fixes the Stub type to have the same semantics as Conn.