reader can leave read buffer in an inconsistent state
Upon read timeout, a conn is declared usable if it didn't read anything
from the underlying net.Conn. However, it is possible that the reader
only consumed data from the read buffer (`br`) and never pulled in data
from the underlying net.Conn. In this case the connection is also in an
inconsistent state and must be closed.