Fix cluster topology parsing for new CLUSTER SLOTS return
As of redis 7.0 the CLUSTER SLOTS command returns an additional element
at the end of each cluster host array. This commit updates the topology
parsing to account for this (by discarding it, for now).

Additionally, this fixes marshaling of the cluster topology, it was
previously marshaling the port as a string when it should be an int.