Fix pool bug involving blocking commands whose context times out
When a blocking command is performed by the pool, it is removed from the
pool during that time so that other commands don't try to share it (and
therefore become blocked themselves). However if the context for the
blocking command was canceled then we were returning the still-blocked
conn to the pool, resulting in subsequent commands being blocked.

Fixes #344